A lot of bugs were fixed; A check for the intersection of orders with an opposite side of the order book has been added. Also possibility to execute such orders as market orders was added.

+++ cancelorder.cpp
@@ -2,16 +2,16 @@
 #define CANCAL_ORDER
 class CancelOrder{
 public:
-    CancelOrder(int id_, long long int time_){
+    CancelOrder(int id_, int64_t time_){
         id = id_;
         time = time_;
     }
     int get_id(){return id;}
-    long long int get_time(){return time;}
+    int64_t get_time(){return time;}
 
-    void set_time(long long int time) {time = time;}
+    void set_time(int64_t time) {time = time;}
 private:
     int id;
-    long long int time;
+    int64_t time;
 };
 #endif //CANCAL_ORDER
